requirements.txt
discord-py-interactions
python-dotenv
openai
requests